requirements.txt
streamlit==1.12.0
torch==1.13.0
torchvision==0.14.0
transformers==4.20.0
Pillow==9.1.0
numpy==1.23.0
pandas==1.4.2
scikit-learn==1.1.1
opencv-python==4.6.0.66
